[Build] Fixed problem in the build shell scripts

+++ setup/prepare.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
 
-rsync "${WORKSPACE}"/bin/illabuild.jar /var/www/illarion/website/media/java
+if [ -f "${WORKSPACE}"/bin/illabuild.jar ]
+then
+	rsync "${WORKSPACE}"/bin/illabuild.jar /var/www/illarion/website/media/java
+fi
